Install webhooks in main operator deployment
The original reason for splitting the webhook out from the main
deployment (b432ee4) was that the cert-manager deployment was
unreliable. The problem was fixed (ec596ffe) but the webhook remained
separate so that installing cert-manager would be optional.

This change brings the webhook back into the main deployment.
Cert-manager is now required. The functionality provided by the webhook
- profile bindings and profile recordings - are core features of the
operator and should not be optional. Since this operator is still in an
alpha state, relying on a third-party tool should be an acceptable
tradeoff for providing the full breadth of functionality with a
consistent, minimal deployment experience. We can reevaluate how to
manage webhook certs further down the road as this operator gains more
user adoption.