Make profilerecording gathering more resilient
This will retry if there's an error getting the profile recording.

It could be that the profile recording gets collected while the hook is
still writing it. So let's try to be more resilient with this. If the
user notices a lot of errors they'll have to delete the recording
object.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>