tree 5044da9ca04d1e0572db1e12cbf594edcd34ae90 parent 13a898e122cbfc0ead48b21dbf68c4ff0e5785e8 author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1620129975 +0300 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1620132312 -0700
Make profilerecording gathering more resilient

This will retry if there's an error getting the profile recording.

It could be that the profile recording gets collected while the hook is
still writing it. So let's try to be more resilient with this. If the
user notices a lot of errors they'll have to delete the recording
object.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>