Fix log enricher read and automatically add /dev/kmsg device
When using CRI-O, we can utilize the annotation
`io.kubernetes.cri-o.Devices` to correctly mount the device into the
container. This means we do not rely on a symlink any more and can use
the device directly.

The enricher read has been fixed because one read always returns a full
line which makes the offset calculation obsolete.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>