Switch to auditd logs for log enricher
We should rely on the auditd logs for the log enricher since the
audit messages will not appear in `/dev/kmsg` if auditd is running.

We can assume that most distributions ship with auditd enabled, and if
the file does not exist we simply handle the log enricher as no-op.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>