tree f0f4df417af1e13ceb6faca31d4d79624cae91e5 parent f0ce89e10be52081095d29455d40005a54d166d1 author Sascha Grunert <sgrunert@redhat.com> 1623999586 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1624012085 -0700
Switch to auditd logs for log enricher

We should rely on the auditd logs for the log enricher since the
audit messages will not appear in `/dev/kmsg` if auditd is running.

We can assume that most distributions ship with auditd enabled, and if
the file does not exist we simply handle the log enricher as no-op.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>