tree 99144cce42667ec1e1fe3420754e8134329e498d parent 2afff1272b2942300050aa7f5e87226765ad81e7 author Jakub Hrozek <jhrozek@redhat.com> 1627416274 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1627457495 -0700
selinux: Print diagnostic message from selinuxd in event on failure

selinuxd does include a diagnostic message for some time and we even
parse it, it was just unused. Instead of changing the CRD just to
include an error message, we might as well just dump the message in an
event.