tree 2867decd7bd2f3da71e3111087136801fa851c96 parent 5c4ebcfb2db5a73fd3f6d33d4d5cb995c3965665 author Jakub Hrozek <jhrozek@redhat.com> 1627479286 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1633345987 -0700
enricher: extend enricher to send SELinux AVCs to the gRPC server

Adds a stub of the gRPC API to send and receive Avc denials. The API is
not implemented at the moment. Depending on which annotation does the
workload have, either a SELinux or a seccomp line is read and dispatched
to the correct API to be read by a client.