enricher: extend enricher to send SELinux AVCs to the gRPC server
Adds a stub of the gRPC API to send and receive Avc denials. The API is
not implemented at the moment. Depending on which annotation does the
workload have, either a SELinux or a seccomp line is read and dispatched
to the correct API to be read by a client.