tree ebbfa4c53612f6264d96c2ac09b19e3e11027d34 parent 1187f235e84467f7a470e93ce921bb20956497fd author Jakub Hrozek <jhrozek@redhat.com> 1627505139 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1633345987 -0700
webhooks: Extend the recording webhook for SELinux

Adds a new ProfileRecording.Spec.Kind enumeration option
"SelinuxProfile. When selected, an annotation is added and the pod would
run with a special permissive domain that is hardcoded at the moment.
This domain would then be used be the enricher to read the AVC denials.

At the moment, the operator doesn't ship with any permissive profile  by
default, only packages one in examples/profilerecording-selinux-logs.yaml.
At the same time, the operator uses the profile unconditionally.