webhooks: Extend the recording webhook for SELinux
Adds a new ProfileRecording.Spec.Kind enumeration option
"SelinuxProfile. When selected, an annotation is added and the pod would
run with a special permissive domain that is hardcoded at the moment.
This domain would then be used be the enricher to read the AVC denials.

At the moment, the operator doesn't ship with any permissive profile  by
default, only packages one in examples/profilerecording-selinux-logs.yaml.
At the same time, the operator uses the profile unconditionally.