tree bb575cfdd4ca6a4bade51967b2d8dd907405d654 parent 3b8e270e412084e1089dfb8df7302c66787f536b author Jakub Hrozek <jhrozek@redhat.com> 1632229016 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1632848563 -0700
CI: Try using ubuntu/focal64 box as a workaround for failing CI

our CI seems to be hitting

    https://bugs.launchpad.net/cloud-images/+bug/1939580

which causes the shared folder to not be mounted and eventually
manifests as the hack/ci/e2e-ubuntu.sh not found, because it's not
copied over to the shared folder.

which is resolved by upgrading to a recent kernel (5.4.0.84). The
generic/ubuntu64 image we have been using appears to be using 5.4.0.81.

Unfortunately, the official ubuntu boxes only ship for virtualbox, which
means that we can't use them in general. But at least let's use them for
our CI, anyone running locally would probably not run into the issue
with the VB shared folders anyway.