CI: Try using ubuntu/focal64 box as a workaround for failing CI
our CI seems to be hitting

    https://bugs.launchpad.net/cloud-images/+bug/1939580

which causes the shared folder to not be mounted and eventually
manifests as the hack/ci/e2e-ubuntu.sh not found, because it's not
copied over to the shared folder.

which is resolved by upgrading to a recent kernel (5.4.0.84). The
generic/ubuntu64 image we have been using appears to be using 5.4.0.81.

Unfortunately, the official ubuntu boxes only ship for virtualbox, which
means that we can't use them in general. But at least let's use them for
our CI, anyone running locally would probably not run into the issue
with the VB shared folders anyway.