Add bpf based syscall recorder
This change adds a new syscall recorder "bpf" which uses an ebpf module
to record the syscalls based on the mount namespace of the container.
This allows tracking multiple PIDs within a single workload. The bpf
recorder features loading and unloading of the module, which means that
it will automatically shutdown if no recordings are ongoing within the
cluster.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>