tree 7e27143f3b65f4e10f22ed874127939e520534e6 parent 0a3ddf12a79bbc3e8217fbc77442aceb369b736b author Sascha Grunert <sgrunert@redhat.com> 1634723043 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1635628277 -0700
Add bpf based syscall recorder

This change adds a new syscall recorder "bpf" which uses an ebpf module
to record the syscalls based on the mount namespace of the container.
This allows tracking multiple PIDs within a single workload. The bpf
recorder features loading and unloading of the module, which means that
it will automatically shutdown if no recordings are ongoing within the
cluster.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>