tree 55d46137b88b45225fdb15fa7dcb4e0ad02df3ba parent 752fe0af0e3f00941f966c161703039e838ef540 author Jakub Hrozek <jhrozek@redhat.com> 1634823421 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1635492051 -0700
enricher: Add enricher items to the backlog when container IDs can't be found

Adds a new "backlog" cache to the enricher module. This cache is to be
used when we find a relevant line, but can't figure out the details of
the workload yet. In this case, the line would be stored in the backlog
and whatever is in the backlog would be processed when all the info is
collected and a line can be dispatched to the GRPC server.

The backlog cache is limited to 128 lines per PID. If there is more
lines stored, the additional lines are simply skipped. This logic
presumes that if the workload doesn't come up for such a long time, we
should just let the cache entries expire to avoid balooning the memory
usage of the enricher.