enricher: Add enricher items to the backlog when container IDs can't be found
Adds a new "backlog" cache to the enricher module. This cache is to be
used when we find a relevant line, but can't figure out the details of
the workload yet. In this case, the line would be stored in the backlog
and whatever is in the backlog would be processed when all the info is
collected and a line can be dispatched to the GRPC server.

The backlog cache is limited to 128 lines per PID. If there is more
lines stored, the additional lines are simply skipped. This logic
presumes that if the workload doesn't come up for such a long time, we
should just let the cache entries expire to avoid balooning the memory
usage of the enricher.