tree a83c99f2bd66457a63d0da67c108c75b1dc08d83 parent e9db01f0da9b2eea1fdf5cad34fe3e89d647f43f author Jakub Hrozek <jhrozek@redhat.com> 1636468579 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1636500627 -0800
Use the verify-boilerplate script from master

We need a verify-boilerplate script version that understands go 1.17
introduced in this commit:
    https://github.com/kubernetes/repo-infra/commit/9d9355c2d32e221fd9a9795fbe54de65c298e92c
unfortunately this means pulling from master as opposed to a tag as
there has not been a release in repo-infra in some time.

Also adds a bunch of autogenerated files to the ignore list.