tree 7319ebfb9a87c6510ee3ce1c1a5475dacba15adb parent 447c89ac82547b70e7ecbcba1a331bac088e00ff author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1637072846 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1637826874 -0800
SelinuxProfile v1alpha2: Add user friendly and raw options

This introduces a more user-friendly SELinux profile. The intent is to
reduce the knowledge that folks need to have about CIL itself, and
instead focus on a more kube-native representation.

This has the following advantages:

* Easier to validate
* Faster to parse
* Easier to read

Underneath, this will generate CIL files that are consumed by the daemon
and eventually installed on the host.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>