SelinuxProfile v1alpha2: Add user friendly and raw options
This introduces a more user-friendly SELinux profile. The intent is to
reduce the knowledge that folks need to have about CIL itself, and
instead focus on a more kube-native representation.

This has the following advantages:

* Easier to validate
* Faster to parse
* Easier to read

Underneath, this will generate CIL files that are consumed by the daemon
and eventually installed on the host.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>