tree 3eb8f654083a88ab6bee5250d59ebd397dd2fde0 parent ff172b93aca627f6a21841c9264c034353636a66 author Jakub Hrozek <jhrozek@redhat.com> 1637275397 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1638267777 -0800
secprofnodestatus: Link the node profile status with profile using a custom label

We were using the profile name as the value of the spo.x-k8s.io/profile-name
label to be able to list node profile statuses for a profile. This is
user-friendly but would fail in case the profile name was longer than 64
characters as that's the label size limit.

Instead, let's rename the label to spo.x-k8s.io/profile-id. In case the
profile name is shorter than the limit, let's still use the name because
it's user-friendly, but in case the name is longer, let's instead use a
sha1 hash of the name with the profile Kind as a prefix.

Using sha1 is OK in this case, because we're not using it for any crypto
purposes, just to shorten a name.