tree 60210053dcf2d010666364f52391c991951c0fb6 parent 47db0e8c8d353d980f091b9ed1ddf71b32ee0fce author Juan Antonio Osorio Robles <jaosorior@redhat.com> 1637572342 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1637579948 -0800
Allow fetching running spod from within it

currently, the SPOD is a singleton that denotes the DaemonSet instance
running on the nodes to do the SPO work. This is mostly just enforced by
the fact that the profilerecorder will fetch the default SPOD to check
its configuration.

This changes that by adding an environment variable on the running pod
that allows one to fetch the SPOD name from the env variable. There is
now a utility function to help with that.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>