Allow fetching running spod from within it
currently, the SPOD is a singleton that denotes the DaemonSet instance
running on the nodes to do the SPO work. This is mostly just enforced by
the fact that the profilerecorder will fetch the default SPOD to check
its configuration.

This changes that by adding an environment variable on the running pod
that allows one to fetch the SPOD name from the env variable. There is
now a utility function to help with that.

Signed-off-by: Juan Antonio Osorio Robles <jaosorior@redhat.com>