spod: Make /etc/selinux.d an emptyDir instead of hostPath
Since the directory only contains policy definitions, not the actual
policy result it is not needed to be a hostPath. The actual policy is
persisted in /etc/selinux (without the .d).

The original idea behind having a hostPath was to run selinuxd on the
nodes directly to lock down even e.g. the kubelet. But that will not
happen, at least not in the short term, so let's reduce the number of
our hostPath mounts a little.