tree 2430fab89bef591e533e0ab4cfae23493c4864b4 parent 248f47e8a838734070561ed0b86716584b83b1b2 author Jakub Hrozek <jhrozek@redhat.com> 1637843472 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1637857713 -0800
spod: Make /etc/selinux.d an emptyDir instead of hostPath

Since the directory only contains policy definitions, not the actual
policy result it is not needed to be a hostPath. The actual policy is
persisted in /etc/selinux (without the .d).

The original idea behind having a hostPath was to run selinuxd on the
nodes directly to lock down even e.g. the kubelet. But that will not
happen, at least not in the short term, so let's reduce the number of
our hostPath mounts a little.