Retry syscall for PID retrieval in bpf recorder
There is a possible race between the event processor and the profile
retrieval RPC, especially when multiple short living processes being
analyzed. It is possible, that we unlock the pidlock while we still have
to wait for a bunch of processes to be analyzed. We overcome this race
by retrying and therefore minimizing the time-gap between the locks.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>