tree f3221736d9f40d7814bd8a36f660c4be3d33aa40 parent df29ce71baa576cc7453b803c125a7dea81c92e3 author Sascha Grunert <sgrunert@redhat.com> 1637924762 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1637933451 -0800
Retry syscall for PID retrieval in bpf recorder

There is a possible race between the event processor and the profile
retrieval RPC, especially when multiple short living processes being
analyzed. It is possible, that we unlock the pidlock while we still have
to wait for a bunch of processes to be analyzed. We overcome this race
by retrying and therefore minimizing the time-gap between the locks.

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>