make: Add PROJECT, CSV and other tooling to create bundles
Uses operator-sdk to create bundles based on existing manifests. By
default uses a single channel called "stable".

To build the bundle image run:
    make bundle; make bundle-build
basic validation is ran automatically as part of the build. To run the
full scorecard, run:
    ./build/operator-sdk scorecard -w 300s ./bundle

The PROJECT file has to be maintained from now on, keeping the list of
APIs as the SDK fills the APIs from PROJECT to the CSV scaffolding at
deploy/base/clusterserviceversion.yaml