tree eab26a0c3b78ef006577bd4e5d6dbe62130524ed parent 178d6c18f75072aa308d5c1cbaa00c8df2ca823f author Jakub Hrozek <jhrozek@redhat.com> 1646923894 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1648453283 -0700
make: install and use opm to create catalogs

Adds Makefile targets to create bundle catalogs. To learn more about
bundle catalogs, visit:
    https://olm.operatorframework.io/docs/tasks/creating-a-catalog/

Note that when building the catalog, the bundle image to be included in
the catalog MUST be reachable via a registry, e.g.:
    BUNDLE_IMGS=quay.io/jhrozek/security-profiles-operator-bundle:v0.4.2-dev-1 \
        make catalog-build

This is an OPM limitation.

Also adds version of packaging tools in dependencies.yaml