tree 964ce451da45f7094f8c58ea974804ce9cb69e97 parent 131206406fcd7f782ea9c99b18987a338827c6db author Cosmin Cojocar <gcojocar@adobe.com> 1646928753 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1647256220 -0700
fix: health probe of seccomp controller should be ready if a node supports seccomp

This fix avoids an endless restart of seccomp container when there isn't any seccompprofile
defined into the cluster. In this situation the reconcile loop never gets executed, and the
seccomp container keeps being restarted.

This is currently happening when the operator is initially installed.

Signed-off-by: Cosmin Cojocar <gcojocar@adobe.com>