examples: OLM manifests
To install OLM:
    export olm_release=0.20.0
    kubectl apply -f https://github.com/operator-framework/operator-lifecycle-manager/releases/download/v${olm_release}/crds.yaml
    kubectl apply -f https://github.com/operator-framework/operator-lifecycle-manager/releases/download/v${olm_release}/olm.yaml

Now on to installing the operator...

Then install the catalogsource first, make sure it reports ready.
    kubectl get catalogsource security-profiles-operator -nolm -oyaml

Then create the namespace, the operatorGroup and finally the
subscription. Creating those should create a CSV that should eventually
transition to Installed:
    kubectl get ip,csv,sub -nsecurity-profiles-operator

Note that the cert-manager dependency still must be installed manually.