manager: create the webhook CR if it doesn't exist.
operator bundles don't allow MutatingWebhookConfiguration as part of the
bundle manifests as a CR (the bundle creation would fail with an error).

Normally it is expected that projects using operator-sdk will manage the
webhooks using OLM and the webhook would be defined in the CSV. That
doesn't work for the current SPO architecture because:
     - the webhook names are auto-generated, just with a common prefix
       That doesn't play well with SPO currently modifying the webhook
       configuration from the code.
     - OLM then manages the certs unconditionally, but that would make
       it harder to not use OLM while packaging and use cert-manager
       instead

All in all, it is easiest to just allow SPO to create the
MutatingWebhookConfiguration. The RBAC permissions were there after all.