Make webhooks' failurePolicy and namespaceSelector configurable
By default, our webhooks target all namespaces. While from a security
standpointm this is OK, because the webhooks react to our CRs and
therefore an explicit RBAC rule must be added to manage them, there is
still a risk that if there is a bug in the webhooks, the function of
critical namespaces might be affected.

This patch exposes the failureMode and the namespaceSelector as
configurable through a newly added webhookOptions object in the spod CR
to enable admins to select which namespaces should be watched by the
webhook and what should happen on failure.

In addition, namespace runlevels 0 and 1 are ignored by defualt on OCP.

Signed-off-by: Jakub Hrozek <jhrozek@redhat.com>