docs: Use kubectl logs --selector name=spod instead kubectl logs ds/spod
kubectl logs ds/spod chooses one of the DS pods at random which might
not "hit" the pod we are looking for. In contrast, the --selector
name=spod invocation logs all the DS pods.

I haven't changed the examples where we follow the logs, because we
might hit the concurrency limit which is set to 5 by default.