tree 62d229e353eacf3e6b8431b7cce97468c0e6f4b7 parent 1e1362a11a1e7a7c9881c93030ae5b70cc70713c author Jakub Hrozek <jhrozek@redhat.com> 1650912320 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1651487713 -0700
deploy: Make kube-rbac-proxy image configurable

Downstreams (in particular OpenShift) might not allow an image from a
third-party container registry to be downloaded, but instead have their
own forks and mirrors.

To be able to point to an OpenShift-approved version of the rbac-proxy,
we need to make the RBAC-proy container image configurable with an
environment variable.