deploy: Make kube-rbac-proxy image configurable
Downstreams (in particular OpenShift) might not allow an image from a
third-party container registry to be downloaded, but instead have their
own forks and mirrors.

To be able to point to an OpenShift-approved version of the rbac-proxy,
we need to make the RBAC-proy container image configurable with an
environment variable.