tree 1db5244a66c8d6cdee4a8ccf216765f649be59a0 parent 503021bdf3c3d256f3ef4644ff635127d8af8d80 author Jakub Hrozek <jhrozek@redhat.com> 1651580114 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1652087479 -0700
spod: Provide explicit selector to the servicemonitor

It was reported (https://issues.redhat.com/browse/CMP-1310) that an
explicit selector in the servicemonitor allows the metrics to show up in
the UI even if the user workload ConfigMap is enabled before SPO
install.

Even though I couldn't reproduce the issue myself, I thought it would be
a nice cleanup.