spod: Provide explicit selector to the servicemonitor
It was reported (https://issues.redhat.com/browse/CMP-1310) that an
explicit selector in the servicemonitor allows the metrics to show up in
the UI even if the user workload ConfigMap is enabled before SPO
install.

Even though I couldn't reproduce the issue myself, I thought it would be
a nice cleanup.