tree 4a8738dd98a17a74368488ad862b92ea701155bc parent 57d564ac71b5d178d8be97951a3b945280e8a5f3 author Jakub Hrozek <jhrozek@redhat.com> 1653051426 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1653367990 -0700
deploy: Annotate the NS to allow running privileged

On some distributions, like OCP 4.11, where PSA is enabled and the
restricted PSA is the default, pods must either run with a very limited
privileges or the namespace must be annotated to allow privileged pods.

Because SPO can't run with such limited privileges (for example selinuxd
or the non-root-enabler must run as root), we need to take the second
option and annotate the namespace.