deploy: Annotate the NS to allow running privileged
On some distributions, like OCP 4.11, where PSA is enabled and the
restricted PSA is the default, pods must either run with a very limited
privileges or the namespace must be annotated to allow privileged pods.

Because SPO can't run with such limited privileges (for example selinuxd
or the non-root-enabler must run as root), we need to take the second
option and annotate the namespace.