tree 150e57e1d71ae643899af2e55090b8259f34fef4 parent 6dbf6982feca8e95ad4e3be26e935765764b5ac6 author Jakub Hrozek <jhrozek@redhat.com> 1653905304 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1654079584 -0700
ocp4: Run deployments with an SA

When SPO is recording a SELinux policy, a special SELinux context is
injected into the workload to trigger auditing all AVC denials in
permissive mode. Using a custom SELinux context is a privileged
operation that can be used only when the SA the workload runs as has the
permissions to use the privileged SCC.

Even though the test is executed normally as cluster-admin, when a
replication controller (e.g. a deployment) is recorded, the pods managed
by the replication controller would run unprivileged by default.

This patch adds several manifests under test/ that contain the SA and
bind it to a role that allows the privileged SCC. This SA is then used
in the deployment, which allows the pods under recording to use the
custom recording SELinux context.