tree 5069ad4e10c0cb390b06aab427a1a6b9e7131db0 parent 4497d79dfa182bf678c5234fbf5789c6b83b0fb0 author Jakub Hrozek <jhrozek@redhat.com> 1654602400 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1654773822 -0700
tests: Explicitly set the context to the SPO namespace before creating the recording SA/role/rolebinding on OCP

This is a workaround for how the OCP's ci-operator work. There's a
namespace created when the tests start which is then removed, but
apparently the context is still set there. Creating resources then fails
with:
    error when creating "test/recording_sa.yaml": namespaces
    "ci-op-hq1cv14k" not found