tree 5b3585814007eab36ea2d3f70dd558894b3aa71e parent e645b213a683ad357d19fe27706e9ce18a30d002 author Jakub Hrozek <jhrozek@redhat.com> 1656616270 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1660290646 -0700
tests: Add the recording RBAC roles and SA dynamically, re-add on every operator redeploy

No idea why we didn't catch this before when running the RBAC roles and
SA on OCP before, but redeploying the operator wouldn't re-add the RBAC
role, role binding and the SA required to grant the ability to use the
SCC needed to record replicated controllers.

This patch adds the needed manifests as part of deploying the operator.

Note that most of the code is only executed on OCP, on other distros,
only the SA is added so that we can always reference the SA regardless
of the distribution.