tests: Add the recording RBAC roles and SA dynamically, re-add on every operator redeploy
No idea why we didn't catch this before when running the RBAC roles and
SA on OCP before, but redeploying the operator wouldn't re-add the RBAC
role, role binding and the SA required to grant the ability to use the
SCC needed to record replicated controllers.

This patch adds the needed manifests as part of deploying the operator.

Note that most of the code is only executed on OCP, on other distros,
only the SA is added so that we can always reference the SA regardless
of the distribution.