tree b17c313bbfc94e0ecf63626690a26645a857bf1a parent e59bacb8eaa411a8981acfbb41da11214f016c86 author Jakub Hrozek <jhrozek@redhat.com> 1656617525 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1660290646 -0700
tests: Modify the example pod if running on OCP

The test pod in testCaseRunPod uses a container running as root. This
doesn't work by default on OCP, where the security defaults force all
pods to run as non-root regardless of what their Dockerfile says.

As pointed out in another PR that tried to use a non-root container, there
is value in testing a root container as well. To solve this, this patch
adds an openshift-specific annotation, forces the pod to run as root and
in order for all this to be allowed, sets the SA that binds to a role
that allows using privileged SCCs.

Tested on OCP 4.11.