tests: Modify the example pod if running on OCP
The test pod in testCaseRunPod uses a container running as root. This
doesn't work by default on OCP, where the security defaults force all
pods to run as non-root regardless of what their Dockerfile says.

As pointed out in another PR that tried to use a non-root container, there
is value in testing a root container as well. To solve this, this patch
adds an openshift-specific annotation, forces the pod to run as root and
in order for all this to be allowed, sets the SA that binds to a role
that allows using privileged SCCs.

Tested on OCP 4.11.