helm: Generate chart using Make and kustomize.
Feedback for PR 1013 raised concerns over maintaining both
a kustomize configuration *and* helm chart.

Talking with Ozz, we decided that it made sense to auto-generate
as much of the helm chart as possible. Unfortunately, the tooling
in this area is lacking. Since there is already a fair amount of
existing code-generation going on, Ozz felt it made sense to
extend that functionality.

This commit is based on an experimental commit Ozz made.
It extends the "overlays" pattern to auto-generate helm chart
data using kustomize. This current commit is likely very
broken because:

  - There is an issue generating data for the other
    deployment methods that I do not understand yet
    (refer to the diffs for "namespace-operator.yaml",
    "operator.yaml", and "openshift-dev.yaml")
  - I have not committed the "Deployment" resource
    for the new chart located at "deploy/helm"
    (I think the intention is to commit this manually,
    and add templating to it as needed)

I left the existing helm chart ("charts/...") in place for
comparison with the auto-generated chart. This can be
removed if we decide to go forward with this strategy.