helm: Fix "commonLabels" for "Deployment" resource.
As a part of shuffling deployment files around to generate
helm chart data, I accidentally removed "commonLabels" from
the "Deployment" resource.

I "fixed" this by copy-pasta'ing them back into the relevant
"kustomization.yaml" file. There does not appear to be a way
to load these from a separate file.

This solution is suboptimal because "commonLabels" are now
specified in multiple files:

  - deploy/base-crds/kustomization.yaml
  - deploy/base/kustomization.yaml
  - deploy/kustomize-deployment/kustomization.yaml

If a new label is added, the user must modify the value in
multiple files.