tree ca724bda5eaaf1ef7249f7ef8280b19fe538436f parent 91c244568706d978a41b8e4b8c5973c139107c6e author Stephen Fox Jr <sfoxjr@equinix.com> 1657916602 -0400 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1658942908 -0700
helm: Fix "commonLabels" for "Deployment" resource.

As a part of shuffling deployment files around to generate
helm chart data, I accidentally removed "commonLabels" from
the "Deployment" resource.

I "fixed" this by copy-pasta'ing them back into the relevant
"kustomization.yaml" file. There does not appear to be a way
to load these from a separate file.

This solution is suboptimal because "commonLabels" are now
specified in multiple files:

  - deploy/base-crds/kustomization.yaml
  - deploy/base/kustomization.yaml
  - deploy/kustomize-deployment/kustomization.yaml

If a new label is added, the user must modify the value in
multiple files.