tree f8f1770969e2300c118b5c0994b9fe77548feef8 parent 0748c42c67b86641d6c6da1b652aa5019c1f0e08 author Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com> 1658995382 +0300 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1659613426 -0700
Run metrics e2e tests in single-node environments only

This is meant to remove some of the flakiness in our tests. We make the
assumption that our vanilla tests are running in single-node mode only,
so we take advantage of that fact and set a boolean that indicates that
it's a single-node test environment.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>