tree b12f3cd6565d7d32fdc8382f44c706169abc2913 parent 6c1e856fe4fb830529dc1f04dc7afc7dd6bfe66b author Jakub Hrozek <jhrozek@redhat.com> 1659442745 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1660290646 -0700
tests: OCP: Set the context in the waitForReadyPodsOCP()

Since we now call deployRecordingSaOcp() directly from the tests that
need the SA, we need to move setting the context earlier in the text
execution, before any manifests are created. In order to avoid adding
yet another setup function that would only be used by OCP, let's set the
context in the waitForReadyPodsOCP() function which was previously
empty.