tests: OCP: Set the context in the waitForReadyPodsOCP()
Since we now call deployRecordingSaOcp() directly from the tests that
need the SA, we need to move setting the context earlier in the text
execution, before any manifests are created. In order to avoid adding
yet another setup function that would only be used by OCP, let's set the
context in the waitForReadyPodsOCP() function which was previously
empty.