tree 4df8c41cca4a77f5deea1757a35c80c273e5b62c parent 3d7d017d9724fd67ad844df904d51a5ba1ebfccd author Stephen Fox Jr <sfoxjr@equinix.com> 1659555259 -0400 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1659601547 -0700
helm: Package and save chart in a GitHub release.

This adds a GitHub Actions workflow that runs "helm package"
on the helm chart and saves it as a workflow artifact. If the
workflow was triggered by a GitHub release being published,
it automatically adds the package to the release that triggered
the workflow.

Note: The package is only attached to a release when it is
published. A draft release will not trigger the workflow.
Refer to the GitHub documentation for more information:
https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#release

The reason for running "helm package" on both pull requests
and release-publishings is to ensure a change introduced in
a PR does not break the workflow automation (it would be very
unfortunate to discover that the workflow broke when we create
a GitHub release).