tree 577854e15c575024d8ad946ffe14c2b508ff7b08 parent edde4d0f3aa73952168e08f84ea4582b52234a91 author Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com> 1660562543 +0300 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1660917240 -0700
Enhance container builds

This uses docker instead of podman/buildah. With this, it adds extra
metadata to the container images. Finally, this introduces Trivy
to do vulnerability scans after building
images.

These scans can alert us to know when a tag bump is needed or when we're
misconfiguring an image.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>