tree e5f3a77a8c2277dc988ba7318b510ba5453943f0 parent 98edee52c4b3538de6e9194e71838bcd2d5d6f7a author Jakub Hrozek <jhrozek@redhat.com> 1660893618 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1660918791 -0700
deploy: Set the NS for spod SA explicitly to allow privileged pods under OCP 4.12+

OCP 4.12+ is trying to make PSA and SCCs coexist. In order to find out
whether the workload can avoid being restricted by the default
("restricted") PSA profile, there is a label synchronizer that takes a
look at the available RBAC permissions and if there is an SA that is
allowed to use a privilged SCC, then the namespace is allowed to be
labeled with `pod-security.kubernetes.io/enforce: privileged`.
More details can be found at:
    https://github.com/openshift/enhancements/blob/master/enhancements/authentication/pod-security-admission-autolabeling.md

It seems that this mechanism relies on having the namespace set
explicitly for RoleBinding's `.subjects[].namespace` to be set when
`.subjects[].kind` is ServiceAccount even though this is not required by
the RBAC system to successfully bind the SA to a Role

This might be a bug:
    https://issues.redhat.com/browse/OCPBUGS-160
however, setting the namespace explicitly does no harm and lets us run
the pods in a privileged manner on OCP 4.12+