controller: Add a new recording merger controller
Adds a new controller called recordingmerger that watches for
profilerecordings that are about to be deleted. For those recordings,
all partial profiles are merged into a final profile and created.

In order to avoid repeating ourselves in the merge code, let's add a new
interface MergeableProfile with two methods: Get() and Merge() which is
implemented by structures private to the controller that embed
SeccompProfile and SelinuxProfile respectively.

Moves UnionSyscalls from the seccomp controller to a utility module
along with its unit tests.

RBAC changes:
    - adds delete/deletecollection for both seccomp and selinux profiles
      to the SPO controllers in order to remove the partial profiles.