tree 5307d4c41fb95a5bb782c3c165c64fea4a318b32 parent 9c8cd6bc6b275371e8162b13bc7f5e16bcb9bafd author Jakub Hrozek <jhrozek@redhat.com> 1661093002 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1665040552 -0700
controller: Add a new recording merger controller

Adds a new controller called recordingmerger that watches for
profilerecordings that are about to be deleted. For those recordings,
all partial profiles are merged into a final profile and created.

In order to avoid repeating ourselves in the merge code, let's add a new
interface MergeableProfile with two methods: Get() and Merge() which is
implemented by structures private to the controller that embed
SeccompProfile and SelinuxProfile respectively.

Moves UnionSyscalls from the seccomp controller to a utility module
along with its unit tests.

RBAC changes:
    - adds delete/deletecollection for both seccomp and selinux profiles
      to the SPO controllers in order to remove the partial profiles.