tree f364c496388d891ba540bf1ebc5e295ddd222322 parent 1cf3492364ea0a86c42e5efe8bfcd8f8d3ec64b4 author Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com> 1661154407 +0300 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1661162144 -0700
Fix tag usage and image loading in image builds

We cannot both push and load at the same time, so logic was added to
make that work. We also now only do container scanning on the first tag
since we cannot use multiple tags as the output of the metadata-action
would normally provide.

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>