recording webhook: don't use the replica name to track pods being recorded, use labels instead
The recording webhook used to track container replicas in a sync.Map,
but its usage was problematic because:
    - all replicas were deleted in case any of replicas with the same
      generatedName were used. This meant replica numbers used for
      recorded policies were being reused and not all containers would
      end up having a recorded policy
    - if a replica was removed, the profilerecording would lose the
      status and finalizers
    - using the replica tracking in the sync.Map means that the webhook
      is maintaining its state, which it shouldn't. While this patch
      doesn't remove the state completely it reduces its usage.

Instead of the above, this patch maintains the finalizer and the status
as a combination of examining the pod being admitted as well as the
currently existing pods which are listed during admission.

The replica number is then always increasing by one e.g. across scaleups
or scaledowns of replicated pods.

Finally, because the webhook is watching all namespaces, but the
sync.Map was only storing plain names, let's also add namespace as a
prefix for the syncMap. This allows to empty the sync.Map once no pods
for a single recording exist, keeping the memory usage low over time.