tree 2c63467650cbbd693cbf27b68e56020e8aca504e parent 3a329e16e8221354dc927d4e9064c7d3d260c0a4 author Jakub Hrozek <jhrozek@redhat.com> 1662402366 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1662557798 -0700
recording: Remove state from the webhook, use a random nonce instead

Instead of keeping track of the replicas in the webhook, let's use a
random nonce to distinguish between different container replicas. When
the profile is recorded for a replicated container, we already know the
pod name, so let's use the hash of the pod to append to the container.
This would also make it easier to cross-reference the pods and the
containers.

In order to parse the recording annotations easily, we also switch from
using dashes in the recording annotations to underscores. Because
underscores can't be used in k8s object names, they make it easy to
split the annotations along them and are allowed in annotations.