tests: Only check the worker nodes testCaseLongSeccompProfileName
The test testCaseLongSeccompProfileName was written with assumption that
the profile being tested is only deployed on worker nodes, probably
because the upstream test platforms only schedule spod there.

On clusters where spod is scheduled on all nodes (OCP), this assumption
breaks and the test would list node statuses of both controlplane and
worker nodes but compare that with the number of workers only.

Let's list nodestatuses on worker nodes only to work around that issue.