tree 41d869e7892abfb26b8bc7ffc26aca9b0c7174a2 parent 7cd33c276ba3d36ad89ab718fee111b56e882637 author Jakub Hrozek <jhrozek@redhat.com> 1662722424 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1662982884 -0700
tests: Only check the worker nodes testCaseLongSeccompProfileName

The test testCaseLongSeccompProfileName was written with assumption that
the profile being tested is only deployed on worker nodes, probably
because the upstream test platforms only schedule spod there.

On clusters where spod is scheduled on all nodes (OCP), this assumption
breaks and the test would list node statuses of both controlplane and
worker nodes but compare that with the number of workers only.

Let's list nodestatuses on worker nodes only to work around that issue.