nodestatus: Add a new profile status "Partial" for profiles yet to be merged
Adds a new profile status "Partial" to both the nodestatus CRD and by
extension the profiles itself. This status is used to denote profiles
that are not to be used right away, but those that are about to be
merged.

Apart from managing the status itself, the node finalizer would also
remove the finalizer preventing the recording from being deleted if all
the recordings go away. This covers the use case where someone records
partial policies and then just wants to remove them instead of merging.