tree 00cce1e87ac68475a8ae58b5ebf760176ff3fc12 parent 65d28421497bd3de5be7ca5019ca798da653d6e5 author Jakub Hrozek <jhrozek@redhat.com> 1663078977 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1665040552 -0700
nodestatus: Add a new profile status "Partial" for profiles yet to be merged

Adds a new profile status "Partial" to both the nodestatus CRD and by
extension the profiles itself. This status is used to denote profiles
that are not to be used right away, but those that are about to be
merged.

Apart from managing the status itself, the node finalizer would also
remove the finalizer preventing the recording from being deleted if all
the recordings go away. This covers the use case where someone records
partial policies and then just wants to remove them instead of merging.