profilerecording: Create partial profiles if recording sets mergeStrategy=containers
When the mergeStrategy field of profileRecording is set to containers,
the profilerecorder controller in the deamon would create profiles with
the label profilebase.ProfilePartialLabel set to true. The profiles
would furthermore contain a label linking to the profileRecording and
the name of the container which will later aid in reconstructing the
merged profile.

When the partial profiles are created, a finalizer is set on the
profileRecording which prevents its deletion unless the partial profiles
are either deleted manually by the admin or merged.

RBAC changes:
    - the profilerecording controller (running under the spod identity)
      is now allowed to list,watch and get profilerecordings