tree 0a808bfdedaa5f7993d36a5375bc7a5ed6413ad3 parent a026ffc5bda77c13d5da977f31bbd0a409371d68 author Jakub Hrozek <jhrozek@redhat.com> 1663700032 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1665040552 -0700
profilerecording: Create partial profiles if recording sets mergeStrategy=containers

When the mergeStrategy field of profileRecording is set to containers,
the profilerecorder controller in the deamon would create profiles with
the label profilebase.ProfilePartialLabel set to true. The profiles
would furthermore contain a label linking to the profileRecording and
the name of the container which will later aid in reconstructing the
merged profile.

When the partial profiles are created, a finalizer is set on the
profileRecording which prevents its deletion unless the partial profiles
are either deleted manually by the admin or merged.

RBAC changes:
    - the profilerecording controller (running under the spod identity)
      is now allowed to list,watch and get profilerecordings