tree e8e6dfa3daa8c6e7a77283497290c28192042521 parent becfe9ce0a82360b8ca788b9772649fe66e15a81 author Jakub Hrozek <jhrozek@redhat.com> 1664447310 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1665040552 -0700
recording: Error out if the profile recording name is too long

Since we now use the profile recording name as a label, we need to check
its name. Since we don't have a dedicated controller and status for the
recording, we emit an event. Moreover, we error and stop reconciling in
the recording controller.