tree cd23904a0b1ebdadd0889a00aa0dda77d8708dd9 parent 5ea32f6446d76d01c828b435f6d13f9e7c9c2e6d author Jakub Hrozek <jhrozek@redhat.com> 1664999659 +0200 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1665737086 -0700
Make the SPO webhooks opt-in

The SPO webhooks have been cluster-wide by default. This is great for
functionality as both recording and binding just works, but might be
dangerous - if there is a bug in the webhook or the webhook is not
replying, then all namespaces might be affected.

This patch switches the default to opt-in, meaning that only namespaces
labeled with spo.x-k8s.io/enable-binding or
spo.x-k8s.io/enable-recording would be listened on.