tree 3bbb34712ce72427f19f35421be8e893cc37286d parent a8a69089e31b9ef2c7a527c5844f10b224ade7a5 author Jakub Hrozek <jhrozek@redhat.com> 1667741103 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1668090415 -0800
selinux: Add a permissive boolean

Similarly to how seccomp profiles allow setting the default action to
LOG, let's allow the selinux profiles to be set to permissive mode. This
way, all calls would have still be allowed, but would appear in
audit.log.

This is useful in case the user is just iterating on a policy and wants
to deploy a first version of the policy in the wild without causing AVC
denials.