tree b7a2109bb800956e9ee6ff4928fd1702905a6399 parent e6e1063c21fb5aaac879bc7ae496272f6cb7e7d0 author Jakub Hrozek <jhrozek@redhat.com> 1668455920 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1668610849 -0800
make: Add an option to produce a dynamic build

OCP does not compile the binaries it runs with upstream go, but rather a
fork of go that uses OpenSSL for cryptography:
    https://github.com/golang-fips

This fork uses dlopen() on startup to load symbols that correspond to
the crypto functions:
    https://github.com/golang-fips/openssl-fips/blob/b9297ed11fc8890f74822482c46f4c9cb5507251/openssl/openssl.go#L53
    https://github.com/golang-fips/openssl-fips/blob/eeda1baae76c92112220de78c5a498a6ba0abb6b/openssl/goopenssl.h#L66-L72

For some reason, this dlopen initialization does to work well for SPO on
OpenShift clusters running in FIPS mode where the dlopen call just
segfaults:
```
(gdb) frame 0
"libcrypto.so.1.1") at dlopen.c:78
78          return _dlfcn_hook->dlopen (file, mode, DL_CALLER);
(gdb) p _dlfcn_hook
$1 = (struct dlfcn_hook *) 0x0
(gdb) list
73      void *
74      __dlopen (const char *file, int mode DL_CALLER_DECL)
75      {
76      # ifdef SHARED
77        if (!rtld_active ())
78          return _dlfcn_hook->dlopen (file, mode, DL_CALLER);
79      # endif
80
81        struct dlopen_args args;
82        args.file = file;
```

To work around this issue, this patch adds the option to compile SPO as
a dynamically linked executable.

Please see https://issues.redhat.com/browse/OCPBUGS-3431 for more
details.