tree 26199995cd8d81607daa3c6d7f04ebe7ddb5eb62 parent 424a6b5ba229d09b4c76bed1aa0fd813d48b15f2 author Jakub Hrozek <jhrozek@redhat.com> 1669065289 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1669123816 -0800
workloadannotator: Prevent crash with non-operator seccomp profiles set in annotation

We would check if the seccomp profile set in the pod or its containers
are owned by the operator, but we did not do the same check for profiles
coming through the annotation. If such annotation existed, the workload
annotator would have attempted to parse it, the returned slice would not
have the expected number of elements which would crash the workload
annotator.