workloadannotator: Prevent crash with non-operator seccomp profiles set in annotation
We would check if the seccomp profile set in the pod or its containers
are owned by the operator, but we did not do the same check for profiles
coming through the annotation. If such annotation existed, the workload
annotator would have attempted to parse it, the returned slice would not
have the expected number of elements which would crash the workload
annotator.