tree 53ba0e21b5fa42e280262a4070d11be41806766d parent 7ec66ffc59a281b8740462284f2a1d94fcb5f055 author Jakub Hrozek <jhrozek@redhat.com> 1669129386 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1669711162 -0800
bundle: Support SingleNamespace and AllNamespaces install methods

I was wrong when I wrote the initial OLM packaging and didn't really
follow the OLM installation modes
(https://olm.operatorframework.io/docs/advanced-tasks/operator-scoping-with-operatorgroups/)

I incorrectly assumed that the installation mode controls where the
operator is installed, but it is supposed to also pertain to where the
watches are supposed to be set.

Even though the operator is really not supposed to be multi-tenant due
to spod modifying the profiles on the hosts, supporting all installation
methods makes sense from the point of view of scoping the watches.