tree bedbb69fa925998b9623e26f1c98351d99c43bcb parent 8a4e694ff703bf84c28209c30894c1642ef7ea23 author Jakub Hrozek <jhrozek@redhat.com> 1669294643 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1669711162 -0800
bundle: Let OLM manage Service Accounts

This patch enables OLM to handle RBAC rules and SAs based on the
namespace SPO is installed into. Without the patch, the RBAC rules would
always reference SAs in the default operator namespace regardless of
where the operator is really installed. With the patch, OLM dynamically
deploys the SAs to the operator namespace, makes them owned by the CSV
which also means that the SAs and the RBAC rules are removed when the
operator is uninstalled. That is not typically important when the
operator has a dedicated namespace because the namespace would have been
removed during uninstall as well, but is important in case the operator
would be sharing a namespace with something else.

The most visible change to the user is that the operator can be
installed into any namespace through OLM. The following manifest
```
apiVersion: operators.coreos.com/v1alpha2
kind: OperatorGroup
metadata:
  name: spo-single-namespace
  namespace: spo-lives-here
spec:
  targetNamespaces:
  - spo-test
```
Installs SPO to a namespace called `spo-lives-here` and only instructs it to
reconcile changes in the namespace called `spo-test`.