e2e: Improve OLM tests to include installation methods
Instead of just running a simple test that installs whatever defaults we
have in the examples, let's actually test all available OLM modes. For
each mode, assert that the CSV is installed the way we want and the CRs
are being listened to in the namespaces we care about only.