tree 4f32ada385651b9a1303aec14014db7d94fa5662 parent d6e99f21bc854a4fc9944b10cf86965583cfb58c author Jakub Hrozek <jhrozek@redhat.com> 1669323124 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1669711162 -0800
e2e: Improve OLM tests to include installation methods

Instead of just running a simple test that installs whatever defaults we
have in the examples, let's actually test all available OLM modes. For
each mode, assert that the CSV is installed the way we want and the CRs
are being listened to in the namespaces we care about only.