tree c9784fbbb1110ba5a079506dc08d7426b0f0cded parent 6ab3863cc6f23ad28c7a03f2ee48706c4097c5de author Jakub Hrozek <jhrozek@redhat.com> 1669413130 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1679321958 -0700
spod: Enabling log enricher via env var should enable the profilerecorder, too

Most of the time when our users enable the log enricher, it is to record
workloads and the profilerecorder controller's `collectLogProfiles()`
understandably doesn't work without log enricher.

However, when the log enricher was enabled through an environment
variable, the variable was not passed down from the controller to the
spod and the profilerecorder did not check the presence of the variable,
only the spod.spec.enableLogEnricher field.

This patch passes the variable down to spod from the manager and makes
the profilerecorder check both the attribute and the env var.