spod: Enabling log enricher via env var should enable the profilerecorder, too
Most of the time when our users enable the log enricher, it is to record
workloads and the profilerecorder controller's `collectLogProfiles()`
understandably doesn't work without log enricher.

However, when the log enricher was enabled through an environment
variable, the variable was not passed down from the controller to the
spod and the profilerecorder did not check the presence of the variable,
only the spod.spec.enableLogEnricher field.

This patch passes the variable down to spod from the manager and makes
the profilerecorder check both the attribute and the env var.