tree 3ccf4518be6fd1e69bfacfc72daf9a0e1e34a466 parent 63115f789dfedb5ef7b7dc3802eacb145d74209f author Cosmin Cojocar <gcojocar@adobe.com> 1669741133 +0100 committer Paulo Gomes <paulo.gomes.uk@gmail.com> 1669796707 +0000
Configure the Selinux type tag when the AppArmor is disabled regardless of EnableSelinux flag

The SELinux type tag needs to be configured independent of EnableSelinux flag, because the
SELinux can be active on the node regardless if the SELinux feature is enabled or not in the operator.
For instance, on Flatcar Linux SELinux type tag needs to be set to 'unconfined_t' instead of 'spc_t'
even though SELinux is disabled in order to get the containers to start.