Configure the Selinux type tag when the AppArmor is disabled regardless of EnableSelinux flag
The SELinux type tag needs to be configured independent of EnableSelinux flag, because the
SELinux can be active on the node regardless if the SELinux feature is enabled or not in the operator.
For instance, on Flatcar Linux SELinux type tag needs to be set to 'unconfined_t' instead of 'spc_t'
even though SELinux is disabled in order to get the containers to start.