OLM: Add separators between CR examples to make sure all CRDs have an example in the bundle
The bundle normally contains an annotation called `alm-examples` that is
displayed in UIs to guide users in creating CRs. Some of our CRs did not
have an example, because there were no separators between the examples
and OLM had trouble parsing them.

In addition, adds a separator between the examples and the deployment
manifest.