tree 49221cce4cceed90e1d2e1f3959a36c5d0ab883b parent ea9826e9f7debe8a99f92f3b833cbbf984fda192 author Jakub Hrozek <jhrozek@redhat.com> 1669979946 +0100 committer Kubernetes Prow Robot <k8s-ci-robot@users.noreply.github.com> 1670502022 -0800
OLM: Add separators between CR examples to make sure all CRDs have an example in the bundle

The bundle normally contains an annotation called `alm-examples` that is
displayed in UIs to guide users in creating CRs. Some of our CRs did not
have an example, because there were no separators between the examples
and OLM had trouble parsing them.

In addition, adds a separator between the examples and the deployment
manifest.